MAJOR-TIME-FIXES:Time now uses a library
Luxon Library to fix time, since fixing it myself was a losing battle. Decided to just make my life easier and I went in and changed everywhere that I saw that was related to time and used this!
Also added some more debugging and further testing and it all seems good now!